Added places controllers
